Moved command to tola directory. Alias user emails
